Address review comments, split code into more files
